percents calculated
